Extract setup into script
Having the script in the json is tricky to edit, also it seems on
reinstall the config files are not updated.

Instead in the install script checkout the repository and call the
setup.sh.
As the server have a fixed configuration we don't need to rely on any
configuration variables in pterodactyl.