Allow skyblock plugin for the default profile
To be able to use the skyblock plugin the permissions for the skyblock
commands needs to be set. For the skyblock server everyone should be
able to play skyblock. The 'skyblock.command' permission needs be
allowed.

The `Ranks.sqlite` file is only created after server start. To be able
to add the permission the `Ranks.sqlite` is populated from a sql dump
file.